Refactor config into sub package
config/parser.py
config/file_path.py
