+++ azurectl/config/parser.py
@@ -17,7 +17,7 @@ import sys
 
 
 # project
-from azurectl_exceptions import (
+from ..azurectl_exceptions import (
     AzureConfigAccountNotFound,
     AzureConfigRegionNotFound,
     AzureAccountDefaultSectionNotFound,
@@ -28,7 +28,7 @@ from azurectl_exceptions import (
     AzureConfigAccountFileNotFound,
     AzureConfigDefaultLinkError
 )
-from config_file_path import ConfigFilePath
+from .file_path import ConfigFilePath
 
 
 class Config(object):
@@ -48,7 +48,7 @@ class Config(object):
         filename=None,
         platform=PLATFORM
     ):
-        from logger import log
+        from ..logger import log
 
         self.storage_container_name = storage_container_name
         self.storage_account_name = storage_account_name
