+++ test/unit/cloud_service_test.py
@@ -6,7 +6,7 @@ from mock import patch
 from test_helper import *
 
 from azurectl.account.service import AzureAccount
-from azurectl.config import Config
+from azurectl.config.parser import Config
 from azurectl.azurectl_exceptions import *
 from azurectl.cloud_service import CloudService
