+++ test/unit/commands_compute_shell_test.py
@@ -9,7 +9,7 @@ from test_helper import *
 import azurectl
 from azurectl.account.service import AzureAccount
 from azurectl.commands.compute_shell import ComputeShellTask
-from azurectl.config import Config
+from azurectl.config.parser import Config
 
 
 class TestComputeShellTask:
